backend/requirements.txt
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0
pypdf>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

